organize tests, update missing test, delete old test files
